Fix jenkins failure 1/3
[Tracking bug 23893]

[git-p4: depot-paths = "//depot/software/bioinformatics/tools/GenomicConsensus/": change = 129471]

+++ GenomicConsensus/quiver/diploid.py
@@ -90,7 +90,7 @@ def scoresForPosition(mms, pos):
     noMutScore = [0] * mms.NumReads()
     mutScores_ = [ mms.Scores(mut)
                    for mut in muts ]
-    mutScores = np.column_stack([noMutScore] + mutScores_)
+    mutScores = np.column_stack([noMutScore] + mutScores_).astype(np.float32)
     return mutScores
